Remove broken shutdown hook.
To fix #1328

When System.exit(0) called, shutdown hook will be run. There is no
guarantee in this case GC will collect everything. So a dead lock will
happen here.

Remove this part of code since the core side can ensure the integrity of
db file when other resource is not released.