Adding jni layer interop code to utilize find method.
Testing

Adding header file

Checking to ensure we're on the right Realm before performing the contains check.

Adding changlog

Adding table view code and changing return types to jlong so we can use them in the indexOf

Adding multi-realm test

Removing code that should not have been in the commit

Changing LinkView.cpp method to use commonly used Macro and udpated null checks.

Changing LinkView.cpp method to use commonly used Macro and udpated null checks and moved contains method to override instead of overload.

Using instance of instead of assignable from and using long value instead of returning a boolean.

Checking for row and index validity in the cpp layer. Also removing the not found constant in order to the use the constant in TableOrView

Checking for managed mode. Falling back to default impl. Updating changelog.

Reverting back to to_jlong_or_not_found

Fixing a few compilation issues as well as applying code review updates.

Fixing managed mode bug that I accidentally introduced.

Updates as per @beeender recommended, issues listed below.
After changing cpp code to return -1 tests started failing - items were not found. I'm on the fence if this is expected or not. I'm not to familiar wit the underlying core to know if -1 should be needed.

Also added tests to check to see if a query or list contains an item that is in the same Realm, but from another query/result. This test creates a native crash. I tried to identify the root cause by pulling the tombstone and performing `ndk-stack` and `addr2line` on it but I could not properly identify the symbols and where they might be.

Inspiration and HOWTO to debug NDK from here: http://bytesthink.com/blog/?p=133

Adding nativePtr to JNI call. Also added isLoaded() call. Not contained test failing on out of bounds exception on row indexes.

Changing TableOrView#find() name to TableOrView#sourceRowIndex() as per recommendation of @beeender

Adding check for parent table and re-enabled the other jni targets and fixed incorrect test.
Received help with this from @beeender and @cmelchior.