Add a check whether transformer was applied or not.
This commit adds default implementation of `transformerApplied()` to `RealmProxyMediator` class which returns `false`.
And then Realm transformer adds overriding method to its subclasses and those methods returns `true`.

If Transformer was applied, `RealmProxyMediator.transformerApplied()` returns `true` and
if not, `RealmProxyMediator.transformerApplied()` returns `false`.
That enable us to detect Realm transformer was applied or not at runtime.