Optimistic opening of a Realm file (#3013)
We have anecdotal evidence that multiple processes might be open for the same app. We have not been able to verify this, but the result matches the "IncompatibleLockFile" errors reported in #2459. This happens if the two processes have two versions of Realm (e.g. during app upgrades).

This PR is based on our assumption that such an overlap is not intentional and only happens because one process is started before the previous was completely shut down.

So this PR introduces an optimistic opening scheme where we retry for 3 second before crashing as before.