Integrate Object Store [PART2] - SharedRealm (#3031)
This simplified our code base a lot. Basically all APIs we need from
SharedGroup/Group are wrapped in the SharedRealm. So we can just remove
those classes.

But we do need a few APIs from SharedGroup which is not supplied by
SharedRealm because of async queries. Currently we expose those in a
friend class of ShareRealm, see realm/realm-object-store#141

We are still managing Realm caches in Java although there are mechanism
in OS to do the same thing. The major reason is we have method like
deleteRealm needs information from cache to check if all Realm instances
are closed. The ShareRealm is actually a std::shared_ptr. We hold
the pointer to the std::shared_ptr in Java.
Another fundamental change is that we used to have separated
SharedGroups for DynamicRealm and typed Realm in the same thread. But
now they are using different SharedRealm but actually the different
SharedRealms are point to the same SharedGroup.

And some other code cleanup.