Remove deprecated constructor + add directory() (#3357)
This commit simplifies the RealmConfiguration constructors and also ensures that we always have an Android context. 

It does so by now only having the`RealmConfiguration.Builder(context)` constructor. Custom file locations are now supported through the `directory()` builder method. This also made it possible to simply `assetFile(Context, location)` to only `assetFile(location)`.

Having the Context means that we are now able to access system services and other framework classes without exposing any Android functionality in any potential interface (which will be needed to support Realm on the JVM).