DeleteLocalRef when the ref is created in loop (#3366)
Add wrapper class for JNI local reference to delete the local ref after
using it.

This is reported by user on helpscout:
https://secure.helpscout.net/conversation/244053233/6163/?folderId=366141

And some useful explanation can be found:
http://stackoverflow.com/questions/24289724/jni-deletelocalref-clarification

Normally the local ref doesn't have to be deleted since they will be
cleaned up when program returns to Java from native code. Using it in a
loop is obvious a corner case: the size of local ref table is relatively
small (512 on Android). To avoid it, the local ref should be deleted
when using it in a loop.