Fix LinkView based RealmResults's tests behaviour
The original tests for deleted LinkView based RealmResults are
incorrect. Before the next event loop, even if the LinkView is not valid
anymore, for the purpose of maintaining the stable iterator, the
RealmResults has to stay the same without calling sync_if_needed. The
results becomes empty only if the we enter into the next event loop and
reattach to the original TableView.

Also, a core bug was found during changing the tests, see
https://github.com/realm/realm-core/issues/2378