Fix type based tests
The original tests rely on the timing of the global listener and type
based listeners which are changed a lot by integration of OS async.
So rewrite those tests by removing the dependencies of global listener
without changing the testing purpose.

PS. : The global listener won't be called if it is added after the local
transaction committed.