Use RealmNotifier for RealmObject listener
Ideally we should use the object notifications from object store, but
since it is still in progress and the KVO notifications is not generic
enough for us to use (there are some logic issues), we simply use the
RealmNotifier to trigger the RealmObject listeners. There will be false
positive notifications with this implementation since we are only
checking the Row's table version. But this problem exists even before
this commit.