Remove outdated addSearchIndexTwice
RealmObjectSchema.addIndex() is not idempotent. The test was created
in #1544 for issue #1385 which was using Table.addSearchIndex().
RealmObjectSchema.addIndex() will just throw when called twice, it is
guaranteed by test addIndexFieldModifier_alreadyIndexedThrows.

The test can accidentally pass is just because of
default-before-migration.realm was created before 0.82 when the primary
key was not auto-indexed yet.