copyToRealmOrUpdate cause dup values in list field (#5023)
Close #4957
We are using HashMap to maintain a cache of added managed RealmObject
when doing copyOrUpdate. But if the given object graph contains objects
which have same PK but are different java object, cache missing will
happen. That caused the object got set multiple times. For non-list
fields, it is not a problem since it will eventually has the expected
value if user's objects graph is consistent. For list, that results the
elements in the list are added multiple times since we don't clear the
RealmList in copy().