Enable downloadProgressListener_indefinitely
In createRemoteData we try to identify the uploading is finished by two
conditions:

- Check if the newly created data exists in the Realm on the callback thread.
- Check if uploadable == uploaded

This works if the callback is called on the sync client thread.
But actually there is a chance the callback gets called immediately after
registered because of some other data created before just got uploaded. In
this case, the data does exist in the Realm on the caller thread (we just
add on the same thread), and the uploaded could equal to the uploadable.

By only accept the callback on the sync client thread by if (threadId ==
Thread.currentThread().getId()) will make sure the data has been uploaded.